refactor: move custom user options into options.py
